LayoutTests/fast/canvas/webgl/renderer-and-vendor-strings.html

<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>
description("Verifies the contents of the RENDERER and VENDOR strings to avoid regressions.")

if (window.initNonKhronosFramework)
    window.initNonKhronosFramework(false);

var gl = create3DContext();

// Consensus in the WebGL working group has been to mask the
// underlying hardware's RENDERER and VENDOR strings, which leak a
// certain amount of personally identifiable information. This test is
// intended only to catch accidental regressions, not to enforce the
// specific strings.

shouldBe("gl.getParameter(gl.RENDERER)", '"WebKit WebGL"');
shouldBe("gl.getParameter(gl.VENDOR)", '"WebKit"');
</script>

<script src="../../js/resources/js-test-post.js"></script>
</body>
