LayoutTests/fast/canvas/webgl/renderer-and-vendor-strings-expected.txt

Verifies the contents of the RENDERER and VENDOR strings to avoid regressions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gl.getParameter(gl.RENDERER) is "CyberKit WebGL"
PASS gl.getParameter(gl.VENDOR) is "CyberKit"
PASS successfullyParsed is true

TEST COMPLETE
